Move all cardinality aspects under impl to draw the line between exposed API and internal impl
