+++ hazelcast/src/main/java/com/hazelcast/cardinality/impl/operations/ReplicationOperation.java
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-package com.hazelcast.cardinality.operations;
+package com.hazelcast.cardinality.impl.operations;
 
-import com.hazelcast.cardinality.CardinalityEstimatorContainer;
-import com.hazelcast.cardinality.CardinalityEstimatorDataSerializerHook;
-import com.hazelcast.cardinality.CardinalityEstimatorService;
+import com.hazelcast.cardinality.impl.CardinalityEstimatorContainer;
+import com.hazelcast.cardinality.impl.CardinalityEstimatorDataSerializerHook;
+import com.hazelcast.cardinality.impl.CardinalityEstimatorService;
 import com.hazelcast.nio.ObjectDataInput;
 import com.hazelcast.nio.ObjectDataOutput;
 import com.hazelcast.nio.serialization.IdentifiedDataSerializable;
