+++ hazelcast/src/main/java/com/hazelcast/client/impl/protocol/task/cardinality/CardinalityEstimatorAggregateMessageTask.java
@@ -16,8 +16,8 @@
 
 package com.hazelcast.client.impl.protocol.task.cardinality;
 
-import com.hazelcast.cardinality.CardinalityEstimatorService;
-import com.hazelcast.cardinality.operations.AggregateOperation;
+import com.hazelcast.cardinality.impl.CardinalityEstimatorService;
+import com.hazelcast.cardinality.impl.operations.AggregateOperation;
 import com.hazelcast.client.impl.protocol.ClientMessage;
 import com.hazelcast.client.impl.protocol.codec.CardinalityEstimatorAggregateCodec;
 import com.hazelcast.client.impl.protocol.task.AbstractPartitionMessageTask;
