+++ hazelcast/src/main/java/com/hazelcast/core/HazelcastInstance.java
@@ -16,6 +16,7 @@
 
 package com.hazelcast.core;
 
+import com.hazelcast.cardinality.CardinalityEstimator;
 import com.hazelcast.config.Config;
 import com.hazelcast.durableexecutor.DurableExecutorService;
 import com.hazelcast.logging.LoggingService;
