+++ hazelcast/src/main/resources/META-INF/services/com.hazelcast.DataSerializerHook
@@ -20,5 +20,5 @@ com.hazelcast.cache.impl.CacheDataSerializerHook
 com.hazelcast.ringbuffer.impl.RingbufferDataSerializerHook
 com.hazelcast.wan.impl.WanDataSerializerHook
 com.hazelcast.query.impl.predicates.PredicateDataSerializerHook
-com.hazelcast.cardinality.CardinalityEstimatorDataSerializerHook
+com.hazelcast.cardinality.impl.CardinalityEstimatorDataSerializerHook
