+++ hazelcast/src/test/java/com/hazelcast/cardinality/CardinalityEstimatorAdvancedTest.java
@@ -1,6 +1,5 @@
 package com.hazelcast.cardinality;
 
-import com.hazelcast.core.CardinalityEstimator;
 import com.hazelcast.core.HazelcastInstance;
 import com.hazelcast.test.HazelcastParallelClassRunner;
 import com.hazelcast.test.HazelcastTestSupport;
