+++ hazelcast/src/test/java/com/hazelcast/cardinality/CardinalityEstimatorBackupTest.java
@@ -16,8 +16,9 @@
 
 package com.hazelcast.cardinality;
 
+import com.hazelcast.cardinality.impl.CardinalityEstimatorContainer;
+import com.hazelcast.cardinality.impl.CardinalityEstimatorService;
 import com.hazelcast.core.HazelcastInstance;
-import com.hazelcast.core.CardinalityEstimator;
 import com.hazelcast.spi.impl.PartitionSpecificRunnable;
 import com.hazelcast.spi.impl.operationservice.impl.OperationServiceImpl;
 import com.hazelcast.test.HazelcastTestSupport;
